Move distributions to its own subpackage.
