+++ pycbc/distributions/uniform_sky.py
@@ -1,0 +1,15 @@
+
+from pycbc.distributions import uniform_angle
+
+class UniformSky(uniform_angle.UniformSolidAngle):
+    """A distribution that is uniform on the sky. This is the same as
+    UniformSolidAngle, except that the polar angle varies from pi/2 (the north
+    pole) to -pi/2 (the south pole) instead of 0 to pi. Also, the default
+    names are "dec" (declination) for the polar angle and "ra" (right
+    ascension) for the azimuthal angle, instead of "theta" and "phi".
+    """
+    name = 'uniform_sky'
+    _polardistcls = uniform_angle.CosAngle
+    _default_polar_angle = 'dec'
+    _default_azimuthal_angle = 'ra'
+
